gcp/gcf/bigquery_export/requirements.txt
google-cloud-bigquery == 2.7.0
google-cloud-storage == 1.35.0
